Add recursion limit to FFI safety lint
Fixes stack overflow in the case of recursive types
